+++ sixtyfps_compiler/generator/cpp.rs
@@ -1432,10 +1432,13 @@ fn compile_expression(
                 unreachable!()
             }
             BuiltinFunction::ColorBrighter => {
-                "[](const auto &color, float factor) {{ return color.brighter(factor); }}".into()
+                "[](const auto &color, float factor) { return color.brighter(factor); }".into()
             }
             BuiltinFunction::ColorDarker => {
-                "[](const auto &color, float factor) {{ return color.darker(factor); }}".into()
+                "[](const auto &color, float factor) { return color.darker(factor); }".into()
+            }
+            BuiltinFunction::ImageSize => {
+                "[](const sixtyfps::Image &img) { return img.size(); }".into()
             }
             BuiltinFunction::Rgb => {
                 "[](int r, int g, int b, float a) {{ return sixtyfps::Color::from_argb_uint8(std::clamp(a * 255., 0., 255.), std::clamp(r, 0, 255), std::clamp(g, 0, 255), std::clamp(b, 0, 255)); }}".into()